requirements.txt
gspread==6.2.1
google-auth==2.43.0
selenium==4.38.0
requests==2.32.5
